requirements.txt
Django==1.11.5
django-bootstrap==0.2.4
django-bootstrap3==9.0.0
django-filter==1.0.4
django-widget-tweaks==1.4.1
gunicorn==19.7.1
olefile==0.44
Pillow==4.3.0
pytz==2017.2
whitenoise==3.3.1
